Version working with microkelvin 0.16.0-rkyv.2

+++ src/ops/store.rs
@@ -17,7 +17,7 @@ impl Get {
     pub fn get(
         env: &Env,
         ofs: u64,
-        len: u32,
+        len: u16,
         buf_ptr: i32,
     ) -> Result<(), VMError> {
         trace!("Executing 'get' host function");
